backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bandit==1.9.2
bcrypt==5.0.0
boolean.py==5.0
build==1.4.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
dparse==0.6.4
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
filelock==3.20.2
flatbuffers==25.12.19
fsspec==2025.12.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
librt==0.7.7
license-expression==30.4.4
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
nltk==3.9.2
numpy==2.4.1
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packageurl-python==0.17.6
packaging==25.0
passlib==1.7.4
pathspec==1.0.2
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.5.1
pluggy==1.6.0
posthog==5.4.0
prometheus_client==0.24.1
protobuf==6.33.2
psutil==7.2.1
py-serializable==2.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==3.2.1
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruff==0.14.14
safety==3.7.0
safety-schemas==0.0.16
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
starlette==0.50.0
stevedore==5.6.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tomli==2.4.0
tomli_w==1.2.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
vulture==2.14
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.1
zipp==3.23.0
